+++ packages/apollo-cache-inmemory/src/inMemoryCache.ts
@@ -64,7 +64,7 @@ export class InMemoryCache extends ApolloCache<NormalizedCache> {
     return this.data;
   }
 
-  public read<T>(query: Cache.ReadOptions): Cache.DiffResult<T> {
+  public read<T>(query: Cache.ReadOptions): T {
     if (query.rootId && typeof this.data[query.rootId] === 'undefined') {
       return null;
     }
@@ -181,7 +181,7 @@ export class InMemoryCache extends ApolloCache<NormalizedCache> {
   public readQuery<QueryType>(
     options: DataProxy.Query,
     optimistic: boolean = false,
-  ): Cache.DiffResult<QueryType> {
+  ): QueryType {
     return this.read({
       query: options.query,
       variables: options.variables,
@@ -192,7 +192,7 @@ export class InMemoryCache extends ApolloCache<NormalizedCache> {
   public readFragment<FragmentType>(
     options: DataProxy.Fragment,
     optimistic: boolean = false,
-  ): Cache.DiffResult<FragmentType> | null {
+  ): FragmentType | null {
     return this.read({
       query: this.transformDocument(
         getFragmentQueryDocument(options.fragment, options.fragmentName),